requirements.txt
flask
flask-cors
numpy==1.24.4
pillow
opencv-python-headless
pandas
matplotlib
pyyaml
requests
ultralytics
seaborn

torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
